+++ frontend/src/entities/ProjectComment/api/comments.ts
@@ -4,14 +4,19 @@ interface PostComment {
   commentId?: number;
   content: string;
 }
+// zustand
+// import useStore from './useStore'
+// import { useTokenStore } from './stores/useTokenStore'
+//
 
+// const token = useStore(useTokenStore, (state) => state.bears)
 const token = localStorage.getItem('user');
 
 export const getComments = (projectId: number) => {
   return axios
     .get(`/api/v1/comments/${projectId}`, {
       headers: {
-        Authorization: `Bearer ${token}}`,
+        authorization: `Bearer ${token}}`,
       },
     })
     .then((response) => response.data.data);
@@ -24,7 +29,7 @@ export const postComment = (comment: PostComment) => {
       { content: comment.content, projectId: comment.projectId },
       {
         headers: {
-          Authorization: `Bearer ${token}}`,
+          authorization: `Bearer ${token}}`,
         },
       },
     )
@@ -35,7 +40,7 @@ export const putComment = (comment: PostComment) => {
   return axios
     .put(
       `/api/v1/comments/${comment.commentId}`,
-      { content: comment.content },
+      { content: comment.content, projectId: comment.projectId },
       {
         headers: {
           authorization: `Bearer ${token}}`,
@@ -47,7 +52,7 @@ export const putComment = (comment: PostComment) => {
 
 export const deleteComment = (commentId: number) => {
   return axios
-    .put(`/api/v1/comments/${commentId}`, {
+    .delete(`/api/v1/comments/${commentId}`, {
       headers: {
         authorization: `Bearer ${token}}`,
       },
