+++ frontend/src/shared/types/comments.ts
@@ -1,23 +1,10 @@
-export interface Reply {
-  parentId: number;
-  id: number;
-  nickname: string;
-  profileImageUrl: string;
-  content: string;
-  createdAt: string;
-  likeCount: number;
-  like: boolean;
-}
-
 export interface Comment {
   commentId: number;
   nickname: string;
   profileImageUrl: string;
   content: string;
   createdAt: string;
-  likeCount: number;
-  like: boolean;
-  replies?: Reply[];
+  authorized: boolean;
 }
 
 export interface CommentData {
